+++ src/screens/StoreDetailScreen/index.tsx
@@ -28,7 +28,6 @@ const StoreDetailScreen = ({
       try {
         setLoadingApi(true);
         const { orderProducts } = await getCartList();
-        console.log('orderProducts', orderProducts?.length);
         if (orderProducts && orderProducts.length > 0) {
           await postCartItem(orderProducts);
         }